% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bufferField.R
\name{bufferField}
\alias{bufferField}
\title{Selecting/filtering experimental data based on shapefile buffering.}
\usage{
bufferField(shape, value = c(-1), field = NULL, cex = 1)
}
\arguments{
\item{shape}{crop/select the object area using this shape as reference.}

\item{value}{referent value to be used as a buffer for  the shapefile (must be numeric and negative).}

\item{field}{object of class SpatialPointsDataFrame. If field=NULL, 
only the new shapefile will be generated.}

\item{cex}{point expansion/size. Please check \code{help("points")}.}
}
\value{
A list with two elements
\itemize{
\item \code{newShape} is the \code{Shape} format \code{SpatialPolygonsDataFrame}.
\item \code{newField} is the field filtered based on the shape.
}
}
\description{
This function allows to filter regions on the original data based on shapefile reduction using different buffer values.
}
\details{
bufferField
}
