% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cropField.R
\name{cropField}
\alias{cropField}
\title{Selecting/cropping experimental field from original data}
\usage{
cropField(field, shape = NULL, nPolygon = 1, nPoint = 4, cex = 1)
}
\arguments{
\item{field}{object of class SpatialPointsDataFrame.}

\item{shape}{crop/select the object area using this shape as reference. If shape=NULL, 
the user will need to draw one or more polygon(s).}

\item{nPolygon}{number of polygons.}

\item{nPoint}{number of points necessary to select field boundaries or area per polygon (4 >= nPoint <= 50).}

\item{cex}{point expansion/size. Please check \code{help("points")}.}
}
\value{
A list with three elements
\itemize{
\item \code{shape} is the \code{Shape} format \code{SpatialPolygonsDataFrame}.
\item \code{cropField} is the original object area without the region/area in the shape.
\item \code{selectedField} is the region/area in the shape.  
}
}
\description{
This function allows to crop or select regions on the original data.
}
\details{
cropField
}
