% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filterField.R
\name{filterField}
\alias{filterField}
\title{Selecting/filtering experimental data based on different criteria.}
\usage{
filterField(
  field,
  shape = NULL,
  trait = NULL,
  value = NULL,
  cropAbove = NULL,
  cex = 1
)
}
\arguments{
\item{field}{object of class SpatialPointsDataFrame.}

\item{shape}{crop/select the object area using this shape as reference. If shape=NULL, 
all field data will be used for filtering.}

\item{trait}{vector with the trait to be used as filter criteria.}

\item{value}{referent value of cropping in the data (must be numeric).}

\item{cropAbove}{if TRUE all values above the 'value' will be selected to the new data.}

\item{cex}{point expansion/size. Please check \code{help("points")}.}
}
\value{
A image format stack.
}
\description{
This function allows to filter regions on the original data.
}
\details{
filterField
}
