% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rasterField.R
\name{rasterField}
\alias{rasterField}
\title{Make a raster from shapefile data}
\usage{
rasterField(field, res = 1, trait = NULL, shape = NULL, cex = 1)
}
\arguments{
\item{field}{SpatVector. The shapefile to be converted to raster.}

\item{res}{resolution to make the raster layer.}

\item{trait}{vector with the traits to be converted to raster.}

\item{shape}{a specific area inside the field to be converted to raster. If shape=NULL, all field data will be converted.}

\item{cex}{point size. Please check \code{help("points")}.}
}
\value{
A raster stack in form SpatRaster.
}
\description{
This function makes raster layers from shapefile data.
}
\details{
rasterField
}
