% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sampleField.R
\name{sampleField}
\alias{sampleField}
\title{Sample data points within specified field.}
\usage{
sampleField(field, shape = NULL, size = 0.1, cex = 1)
}
\arguments{
\item{field}{SpatVector. The field to be sampled.}

\item{shape}{a specific area inside the field to be sampled. If shape=NULL, all field data will be sampled.}

\item{size}{the proportion of original data to be sampled (must be values between 0 and 1).}

\item{cex}{point size. Please check \code{help("points")}.}
}
\value{
the sampled field is returned in form SpatVector.
}
\description{
This function samples a specified SpatVector, reducing the data size.
}
\details{
sampleField
}
