% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sdField.R
\name{sdField}
\alias{sdField}
\title{Filter spatial vector data based on standard deviation.}
\usage{
sdField(field, shape = NULL, trait = NULL, value = 1, cex = 1)
}
\arguments{
\item{field}{SpatVector. The field to be filtered.}

\item{shape}{a specific area inside the field to be filtered. If shape=NULL, all field data will be used for filtering.}

\item{trait}{a vector of the traits to be used as filter criteria.}

\item{value}{threshold value of standard deviation to filter the trait data below (must be numeric).}

\item{cex}{point size. Please check \code{help("points")}.}
}
\value{
A SpatVector filtered by standard deviation.
}
\description{
This function filters the original data based on standard deviation of traits.
}
\details{
sdField
}
