% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_boundaryField.R
\name{boundaryField}
\alias{boundaryField}
\title{Making raster from original data}
\usage{
boundaryField(field, draw = FALSE, col = "red", cex = 1)
}
\arguments{
\item{field}{object of class 'raster'.}

\item{draw}{use TRUE for drawing field boundaries.}

\item{col}{line color.}

\item{cex}{point expansion/size. Please check \code{help("points")}.}
}
\value{
shapefile boundary SpatialPolygonsDataFrame.
}
\description{
This function allows to make shapefile boundary from raster layers.
}
\details{
boundaryField
}
