% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bufferField.R
\name{bufferField}
\alias{bufferField}
\title{Create a buffer shapefile within the target geometry}
\usage{
bufferField(shape, value = c(-1), field = NULL, cex = 1)
}
\arguments{
\item{shape}{SpatVector. The target geometry to be buffered. Object of class.}

\item{value}{value of the buffer - measured in mapunits. Must be numeric and negative.}

\item{field}{SpatVector. If input, field cropped by target geometry included in output. If field=NULL, only the buffer will be generated.}

\item{cex}{point size. Please check \code{help("points")}.}
}
\value{
If field=NULL, the buffer SpatVector is returned. If field provided, list with two elements is returned:
\itemize{
\item \code{newShape} is the buffer SpatVector.
\item \code{newField} is the field cropped using the buffered shape, also a SpatVector.
}
}
\description{
This function produces a buffer within the target geometry by a specified value.
}
\details{
bufferField
}
