% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filterField.R
\name{filterField}
\alias{filterField}
\title{Filter spatial vector data based on value}
\usage{
filterField(
  field,
  shape = NULL,
  trait = NULL,
  value = NULL,
  cropAbove = NULL,
  cex = 1
)
}
\arguments{
\item{field}{SpatVector. The field to be filtered.}

\item{shape}{a specific area inside the field to be filtered. If shape=NULL, all field data will be used for filtering.}

\item{trait}{a vector of the traits to be used as filter criteria.}

\item{value}{threshold value to filter the trait data (must be numeric).}

\item{cropAbove}{if TRUE all data above the 'value' will be retained. if FALSE all data below the 'value' will be retained.}

\item{cex}{point size. Please check \code{help("points")}.}
}
\value{
the filtered SpatVector.
}
\description{
This function filters spatial vector data based on trait value.
}
\details{
filterField
}
