% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rasterField.R
\name{rasterField}
\alias{rasterField}
\title{Making raster from original data}
\usage{
rasterField(field, res = 1, trait = NULL, shape = NULL, cex = 1)
}
\arguments{
\item{field}{object of class SpatialPointsDataFrame.}

\item{res}{resolution to make the raster layer.}

\item{trait}{vector with the trait to be used as filter criteria.}

\item{shape}{crop/select the object area using this shape as reference. If shape=NULL, 
the user will need to draw one or more polygon(s).}

\item{cex}{point expansion/size. Please check \code{help("points")}.}
}
\value{
A image format stack.
}
\description{
This function allows to make raster layers from original data.
}
\details{
rasterField
}
