% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sampleField.R
\name{sampleField}
\alias{sampleField}
\title{Sampling experimental data points.}
\usage{
sampleField(field, shape = NULL, size = 0.1, cex = 1)
}
\arguments{
\item{field}{object of class SpatialPointsDataFrame.}

\item{shape}{crop/select the object area using this shape as reference. If shape=NULL, 
the user will need to draw one or more polygon(s).}

\item{size}{reduction proportion of original data (must be values between 0 and 1).}

\item{cex}{point expansion/size. Please check \code{help("points")}.}
}
\value{
A image format stack.
}
\description{
This function allows to reduce the original data size.
}
\details{
sampleField
}
